state: share more indexer functions for config_entries
